Rename sql_store to provider module
