+++ aquadoggo/src/rpc/methods/publish_entry.rs
@@ -3,7 +3,7 @@
 use jsonrpc_v2::{Data, Params};
 use p2panda_rs::storage_provider::traits::StorageProvider;
 
-use crate::db::sql_store::SqlStorage;
+use crate::db::provider::SqlStorage;
 use crate::errors::StorageProviderResult;
 use crate::rpc::{PublishEntryRequest, PublishEntryResponse};
